DTO types refactor

+++ api/src/main/java/com/juanmorschrott/api/converter/HotelDtoToHotelConverter.java
@@ -6,6 +6,8 @@ import com.juanmorschrott.api.model.Hotel;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class HotelDtoToHotelConverter implements Converter<HotelDto, Hotel> {
 
@@ -16,7 +18,7 @@ public class HotelDtoToHotelConverter implements Converter<HotelDto, Hotel> {
                 .id(source.getId())
                 .name(source.getName())
                 .description(source.getDescription())
-                .price(source.getPrice())
+                .price(BigDecimal.valueOf(source.getPrice()))
                 .build();
     }
